% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp}
\alias{clamp}
\title{Constraint a vector to fit inside bounds}
\usage{
clamp(x, thresh_min = 0.01, thresh_max = 0.99)
}
\arguments{
\item{x}{A vector of numerical values.}

\item{thresh_min}{Lower bound.}

\item{thresh_max}{Upper bound.}
}
\value{
The same vector but whose values are clamped to the given bounds.
}
\description{
Constraint a vector to fit inside bounds
}
\examples{
clamp(c(0.1, 0.4, 0.9), thresh_min = 0.3, thresh_max = 0.8)
}
