% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compute_covariance_multinomial}
\alias{compute_covariance_multinomial}
\title{Covariance matrix of totals from multinomial sampling}
\usage{
compute_covariance_multinomial(x, y, pi_i)
}
\arguments{
\item{x}{A matrix (n x p) representing a vector valued interest variable,
where lines correspond to individual.}

\item{y}{A Matrix (n x q) representing a vector valued interest variable,
where lines correspond to individual.}

\item{pi_i}{A vector (n) of inclusion probabilities of each individual.}
}
\value{
The associated covariance matrix (p x q).
}
\description{
Covariance matrix of totals from multinomial sampling
}
\examples{
x <- matrix(c(1, 2, 1, 2), nrow = 2)
y <- matrix(c(1, 3, 4, 5, 6, 7), nrow = 2)
pi_i <- c(0.5, 0.8)
compute_covariance_multinomial(x, y, pi_i)
}
