% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_noise.R
\name{compute_noise_df}
\alias{compute_noise_df}
\title{Generate a table comparing variance under different levels of noise}
\usage{
compute_noise_df(
  population,
  noises,
  x_names,
  n_iter_true,
  sample_fn_list,
  v_approx_fn_list
)
}
\arguments{
\item{population}{A data frame corresponding to the population.}

\item{noises}{A vector with the varying levels of noise.}

\item{x_names}{A vector of the names of the auxiliary variables used to
balance the sample.}

\item{n_iter_true}{Number of iterations to compute variance by Monte Carlo
estimation.}

\item{sample_fn_list}{A function returning a named list of sampling
functions. The function takes \code{x_names} as an argument.}

\item{v_approx_fn_list}{A named list of variance approximations.}
}
\value{
A table with the computed variance for each level of noise.
}
\description{
Generate a table comparing variance under different levels of noise
}
\examples{
population <- data.frame(y = 1:3, x1 = 2:4, pi_i = c(0.2, 0.3, 0.5))
compute_noise_df(
  population,
  noises = c(0.1, 0.5, 1, 2, 5, 10),
  x_names = "x1",
  n_iter_true = 100L,
  sample_fn_list = \(x_names) list(base = sampler_gen(x_names)),
  v_approx_fn_list = list(v_multi = compute_v_approx_multinomial)
)
}
