% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_relative_bias}
\alias{compute_relative_bias}
\title{Compute relative bias of estimations}
\usage{
compute_relative_bias(v_hats, v_true)
}
\arguments{
\item{v_hats}{A vector containing multiple variance estimators.}

\item{v_true}{True variance.}
}
\value{
A scalar corresponding to the relative bias.
}
\description{
This computation requires producing multiple estimations and knowing the
true value (eventually by a Monte-Carlo approach).
}
\examples{
compute_relative_bias(c(1,1,2,3,2), 2)
}
