% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compute_total}
\alias{compute_total}
\title{Compute total of variable \code{y} of the population}
\usage{
compute_total(population)
}
\arguments{
\item{population}{A data frame corresponding to the population.}
}
\value{
The total.
}
\description{
Compute total of variable \code{y} of the population
}
\examples{
create_population(10) |> compute_total()
}
