% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_statistics.R
\name{compute_v_approx_deville_tille}
\alias{compute_v_approx_deville_tille}
\title{Approximation de Deville-Tillé de la variance d'un tirage équilibré}
\usage{
compute_v_approx_deville_tille(population, x_names, tol = 1e-06)
}
\arguments{
\item{population}{Population.}

\item{x_names}{Vecteur des noms des variables auxiliaires utilisées pour
l'équilibrage.}

\item{tol}{Seuil de tolérance pour déterminer les valeurs propres nulles.}
}
\value{
Une approximation de la variance du total sur la variable \code{y} de la
population.
}
\description{
Approximation de Deville-Tillé de la variance d'un tirage équilibré
}
