% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_monte_carlo.R
\name{compute_v_trues}
\alias{compute_v_trues}
\title{Monte Carlo variance computation of estimators of total}
\usage{
compute_v_trues(population, sample_fn_list, n_iter_true)
}
\arguments{
\item{population}{A data frame corresponding to the population.}

\item{sample_fn_list}{A named list of sampling functions to apply to the
population.}

\item{n_iter_true}{Number of iterations of the Monte Carlo simulation.}
}
\value{
A named vector with the variance associated to each sampling
procedure. The names correspond to the names in the list.
}
\description{
The variance associated to multiple sampling procedure can be applied.
}
\examples{
population <- data.frame(y = 1:4, x1 = 5:8, pi_i = rep(0.2, 4))
compute_true(population, list(base = sampler_gen_base("x1")), 20)
}
