% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_population.R
\name{create_population}
\alias{create_population}
\title{Create a population}
\usage{
create_population(n_tot, noise = 1)
}
\arguments{
\item{n_tot}{Total number of individual.}

\item{noise}{Noise parameter in the generation of the variable of interest.}
}
\value{
A \code{tibble} data frame.
}
\description{
Create a population
}
\examples{
create_population(10, noise = 1)
}
