% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_estimation.R
\name{estimate_total}
\alias{estimate_total}
\title{Estimate total from a sample}
\usage{
estimate_total(sample)
}
\arguments{
\item{sample}{A data frame corresponding to a sample of the population.
It contains:
\itemize{
\item \code{y}: the variable of interest.
\item \code{pi_i}: the inclusion probability.
\item \code{s_i}: the (eventually non-integer) indicator of inclusion.
}}
}
\value{
A scalar corresponding to the estimator of the total.
}
\description{
The sample can have non-integer selection of units, for example as a result
of the flight phase.
}
\examples{
samp <- data.frame(y = 1:5, pi_i = rep(0.5, 5), s_i = c(1, 0, 1, 0.5, 0, 0))
estimate_total(samp)
}
