% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{flight_wr}
\alias{flight_wr}
\title{Flight phase of the with-replacement cube method}
\usage{
flight_wr(X, pik, eps = 1e-11)
}
\arguments{
\item{X}{The balancing matrix.}

\item{pik}{The inclusion probability vector.}

\item{eps}{A threshold for determining if a value is null.}
}
\value{
A sampling vector of multiplicities
}
\description{
Flight phase of the with-replacement cube method
}
