% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_noise.R
\name{get_sample_fn_list}
\alias{get_sample_fn_list}
\title{Get list of sample strategies}
\usage{
get_sample_fn_list(x_names)
}
\arguments{
\item{x_names}{A vector of the names of the auxiliary variables used to
balance the sample.}
}
\value{
A named list with sampling functions.
}
\description{
These are mostly balanced sampling algorithms
}
\examples{
get_sample_fn_list(c("const", "pi_i_aux"))
}
