% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_noise.R
\name{get_v_approx_fn_list}
\alias{get_v_approx_fn_list}
\title{Get list of variance approximations}
\usage{
get_v_approx_fn_list()
}
\value{
A named list with variance approximations.
}
\description{
Get list of variance approximations
}
\examples{
get_v_approx_fn_list()
}
