% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{jump_wr}
\alias{jump_wr}
\title{Jump step in the with-replacement cube method}
\usage{
jump_wr(X, pik, eps = 1e-11)
}
\arguments{
\item{X}{The balancing matrix.}

\item{pik}{The sampling/inclusion probability vector before jumping.}

\item{eps}{A threshold for determining if a value is null.}
}
\value{
The sampling vector after the jump.
}
\description{
Will select a random direction in the kernel of the balancing matrix and move
the current sampling/inclusion probability vector to a new state. Given a
direction, two candidate states are selected, corresponding to the moment
where a unit is unselected.
}
