% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_optimisation.R
\name{max_entropy_p}
\alias{max_entropy_p}
\title{Produce the max-entropy probabilities associated to a jumping step}
\usage{
max_entropy_p(lambda, trace = 0)
}
\arguments{
\item{lambda}{}
}
\description{
Produce the max-entropy probabilities associated to a jumping step
}
\examples{
max_entropy_p(c(1, 2, -3), trace = 1)
}
