% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_monte_carlo.R
\name{mc_estimate_total}
\alias{mc_estimate_total}
\title{Repeated estimations of totals from independent samples}
\usage{
mc_estimate_total(population, sample_fn, n_iter, .progress = TRUE)
}
\arguments{
\item{population}{A data frame corresponding to the population.}

\item{sample_fn}{A sampling procedure as a function to apply to
the population.}
}
\value{
A vector of estimated totals for each generated sample.
}
\description{
This function generates multiple estimators of total obtained from repeated
sampling using the same procedure.
}
\examples{
population <- data.frame(y = 1:4, x1 = 5:8, pi_i = rep(0.2, 4))
mc_estimate_total(population, sampler_gen("x1"), 20)
}
