% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_population.R
\name{pi_gen_beta}
\alias{pi_gen_beta}
\title{Beta distribution probability generator}
\usage{
pi_gen_beta(a = 1, b = 5)
}
\arguments{
\item{a}{Parameter of the beta distribution.}

\item{b}{Parameter of the beta distribution.}
}
\value{
A probability generator.
}
\description{
Beta distribution probability generator
}
\examples{
pi_gen_beta(1, 5)
}
