% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_population.R
\name{pi_gen_prop_size}
\alias{pi_gen_prop_size}
\title{Proportional to variable - probability generator}
\usage{
pi_gen_prop_size(n_sample, x)
}
\arguments{
\item{n_sample}{Number of element in the sample.}

\item{x}{Auxiliary variable to which the probabilities should be
proportional to.}
}
\value{
A probability generator.
}
\description{
Note: this does not depend on n, it ideally should depend on population
Now we will assume that the population is not dynamically changing, so it is
possible to precompute the auxiliary variables.
}
\examples{
# This generator assumes that the population is of size 7
pi_gen_prop_size(2, c(1,2,3,5,2,1,1))
}
