% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_population.R
\name{pi_gen_unif}
\alias{pi_gen_unif}
\title{Uniform distribution probability generator}
\usage{
pi_gen_unif(n_sample)
}
\arguments{
\item{n_sample}{Number of element in the sample.}
}
\value{
A probability generator.
}
\description{
Uniform distribution probability generator
}
\examples{
pi_gen_unif(10)
}
