% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudo_inv}
\alias{pseudo_inv}
\title{Computation of the pseudo-inverse of a matrix}
\usage{
pseudo_inv(x, tol = 1e-06)
}
\arguments{
\item{x}{Matrix to pseudo-invert.}

\item{tol}{Threshold to consider an eigenvalue as null.}
}
\value{
A matrix.
}
\description{
Computation of the pseudo-inverse of a matrix
}
\examples{
pseudo_inv(diag(c(1,2,3,0)), tol = 1e-3)
}
