% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduc}
\alias{reduc}
\title{Recover singular vectors corresponding to non-zero singular values}
\usage{
reduc(X, eps = 1e-11)
}
\arguments{
\item{X}{Matrix to recover singular values.}

\item{eps}{Threshold to consider a singular value as null.}
}
\value{
A reduced matrix.
}
\description{
Recover singular vectors corresponding to non-zero singular values
}
\examples{
reduc(matrix(c(1, 2, 3, 4), nrow = 2), eps = 1e-10)
}
