% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_optimisation.R
\name{sample_max_entropy}
\alias{sample_max_entropy}
\title{Sample an index with max-entropy under martingale constraint}
\usage{
sample_max_entropy(lambda)
}
\arguments{
\item{lambda}{}
}
\description{
Sample an index with max-entropy under martingale constraint
}
\examples{
sample_max_entropy(c(1, 2, -3))
}
