% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{sampler_gen}
\alias{sampler_gen}
\title{Standard balanced sampling using the cube method}
\usage{
sampler_gen(x_names, ...)
}
\arguments{
\item{population}{}
}
\value{
A sample with the same columns as the population, but with an
indicator of selection \code{s_i}.
}
\description{
Standard balanced sampling using the cube method
}
