% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{sampler_gen_base}
\alias{sampler_gen_base}
\title{Standard balanced sampling using the cube method}
\usage{
sampler_gen_base(x_names, ...)
}
\arguments{
\item{x_names}{A vector of the names of the auxiliary variables used to
balance the sample.}

\item{...}{Unused extra arguments.}
}
\value{
A sampling function, which given a population, returns a sample with
the same columns as the population, but with an indicator of selection
named \code{s_i}.
}
\description{
Standard balanced sampling using the cube method
}
