% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{sampler_gen_flight_base}
\alias{sampler_gen_flight_base}
\title{Cube method (flight phase)}
\usage{
sampler_gen_flight_base(x_names, ...)
}
\arguments{
\item{x_names}{A vector of the names of the auxiliary variables used to
balance the sample.}

\item{...}{Unused extra arguments.}
}
\value{
A sampling function, which given a population, returns a sample with
the same columns as the population, but with an indicator of selection
named \code{s_i} that can be between 0 and 1.
}
\description{
Cube method (flight phase)
}
