% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{sampler_gen_fuzzy_custom}
\alias{sampler_gen_fuzzy_custom}
\title{Cube method fuzzy sampling (flight phase)}
\usage{
sampler_gen_fuzzy_custom(x_names, ...)
}
\arguments{
\item{population}{}
}
\value{
A sample with the same columns as the population, but with an
indicator of selection \code{s_i} that can be between 0 and 1.
}
\description{
Cube method fuzzy sampling (flight phase)
}
