% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{sampler_gen_wr_copy}
\alias{sampler_gen_wr_copy}
\title{With-replacement cube method sampling using dimensional increase}
\usage{
sampler_gen_wr_copy(x_names, ...)
}
\arguments{
\item{population}{}
}
\value{
A sample with the same columns as the population, but with an
indicator of selection \code{s_i}.
}
\description{
With-replacement cube method sampling using dimensional increase
}
