% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_population.R
\name{set_inclusion_proba}
\alias{set_inclusion_proba}
\title{Set the inclusion probabilities of the population}
\usage{
set_inclusion_proba(population, pi_gen)
}
\arguments{
\item{population}{A data frame corresponding to the population.}

\item{pi_gen}{A probability generator. This generator must take
the number of individual as a parameter.}
}
\value{
The same data frame with two additionnal columns:
\code{pi_i} and \code{pi_i_aux}.
}
\description{
Description
}
\examples{
create_population(10) |> set_inclusion_proba(pi_gen_unif)
}
