% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_noise.R
\name{write_table_noise_tex}
\alias{write_table_noise_tex}
\title{Write the table of noise}
\usage{
write_table_noise_tex(result_df, fixed_sample_fn_list, v_approx_fn_list)
}
\arguments{
\item{result_df}{Table result of the experiment for varying level of noise.}

\item{fixed_sample_fn_list}{List of sampling functions that where used for
producing the table.}

\item{v_approx_fn_list}{List of approximation functions that where used for
producing the table.}
}
\value{
The path where the table was saved.
}
\description{
The table is written in the LaTeX format.
}
