% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_statistics.R
\name{compute_covariance_multinomial}
\alias{compute_covariance_multinomial}
\title{Calcul de la matrice de variance-covariance de deux totaux de variables
d'intérêt associées à un tirage multinomial}
\usage{
compute_covariance_multinomial(x, y, pi_i)
}
\arguments{
\item{x}{Matrice (n x p) représentant une variable d'intérêt vectorielle, les
lignes correspondant aux individus.}

\item{y}{Matrice (n x q) représentant une variable d'intérêt vectorielle, les
lignes correspondant aux individus.}

\item{pi_i}{Vecteur des probabilités d'inclusion de chaque individu.}
}
\value{
Une matrice (p x q) de variance-covariance
}
\description{
Calcul de la matrice de variance-covariance de deux totaux de variables
d'intérêt associées à un tirage multinomial
}
