% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_statistics.R
\name{compute_total}
\alias{compute_total}
\title{Calcul du total sur la variable \code{y} de la population}
\usage{
compute_total(population)
}
\arguments{
\item{population}{Population}
}
\value{
Le total
}
\description{
Calcul du total sur la variable \code{y} de la population
}
