% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_statistics.R
\name{pseudo_inv}
\alias{pseudo_inv}
\title{Calcul du pseudo-inverse d'une matrice}
\usage{
pseudo_inv(x, tol = 1e-06)
}
\arguments{
\item{x}{Matrice à pseudo-inverser.}

\item{tol}{Seuil de tolérance pour déterminer les valeurs propres nulles.}
}
\value{
Une matrice.
}
\description{
Calcul du pseudo-inverse d'une matrice
}
\examples{
pseudo_inv(diag(c(1,2,3,0)), tol = 1e-3)
}
