% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sample_population.R
\name{sampler_gen_srswor}
\alias{sampler_gen_srswor}
\title{Standard without-replacement sampling}
\usage{
sampler_gen_srswor(...)
}
\arguments{
\item{...}{Unused extra arguments.}
}
\value{
A sampling function, which given a population, returns a sample with
the same columns as the population, but with an indicator of selection
named \code{s_i}.
}
\description{
Standard without-replacement sampling
}
